staffLines export to MusicXML is done

+++ music21/musicxml/fromMxObjects.py
@@ -2090,7 +2090,7 @@ def mxToMeasure(mxMeasure, spannerBundle=None, inputM21=None, lastMeasureInfo=No
                     
                     staffLines = None
                     try:
-                        staffLines = float(mxStaffDetails.staffLines)
+                        staffLines = int(mxStaffDetails.staffLines)
                     except TypeError:
                         staffLines = 5
